+++ assignment-6&7/package.json
@@ -10,6 +10,7 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.4.4",
     "react-scripts": "5.0.1",
+    "react-switch": "^7.0.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
