+++ assignment-6&7/src/App.js
@@ -1,7 +1,9 @@
-import { useEffect, useState } from 'react';
+import { useContext, useEffect, useState } from 'react';
 import { Outlet } from 'react-router-dom';
 import './styles.css';
 import { CardComponent, TitleComponent, SearchBar } from './Components';
+import usernames from './utils/usernames';
+import ThemeContext from './Context/ThemeContext';
 
 const CardContainer = ({ userData }) => {
   return (
@@ -17,21 +19,23 @@ const CardContainer = ({ userData }) => {
   )
 }
 
-const usernames = [
-  "akshaymarch7", 
-  "ap221882", 
-  "nikitaj-57", 
-  "Pujarini", 
-  "aditifarkya",
-  "gavandivya",
-  "nitishnivedan",
-  "Bhallora"
-]
+// const usernames = [
+//   "akshaymarch7", 
+//   "ap221882", 
+//   "nikitaj-57", 
+//   "Pujarini", 
+//   "aditifarkya",
+//   "gavandivya",
+//   "nitishnivedan",
+//   "Bhallora"
+// ]
 
 const BodyComponent = () => {
   // UserData will store the data fetched using Github API
   const [ userData, setUserData ] = useState([]);
   const [ filteredData, setFilteredData ] = useState([]);
+
+  const { theme, setTheme } = useContext(ThemeContext);
 
   const fetchUserDetails = async() => {
     let response = await Promise.all(
@@ -50,7 +54,7 @@ const BodyComponent = () => {
   },[])
   
   return (
-    <div>
+    <div style={{ "backgroundColor" : theme === "light" ? "#778899" : "#1e2126"}}>
       <SearchBar userData={userData} setFilteredData={setFilteredData}/>
       <CardContainer userData={filteredData.length ? filteredData : userData}/>
     </div>
@@ -58,11 +62,12 @@ const BodyComponent = () => {
 }
 
 const AppLayout = () => {
+  const [ theme, setTheme ] = useState("light");
   return (
-    <>
+    <ThemeContext.Provider value={{ theme: theme, setTheme: setTheme }}>
       <TitleComponent/>
       <Outlet />
-    </>
+    </ThemeContext.Provider>
     
   )
 }
