+++ assignment-6&7/src/Components/Title/TitleComponent.js
@@ -1,9 +1,15 @@
 import './style.css';
 import { Link } from 'react-router-dom';
+import { useContext } from 'react';
+import ThemeContext from '../../Context/ThemeContext';
 
 const title = "Brainy Fools";
 
-const TitleComponent = () => (
+
+
+const TitleComponent = () => {
+    const { theme, setTheme } = useContext(ThemeContext);
+    return (
     <nav className='nav-container'> 
         <h1>{title}</h1>
         <div className='ml-auto nav-right'>
@@ -15,7 +21,9 @@ const TitleComponent = () => (
                 <h3 id='about'>About</h3>
             </Link>
         </div> 
-    </nav>     
-)
+        <button onClick={() => setTheme(theme === "dark" ? "light" : "dark")}>{theme}</button>
+    </nav>   
+    )  
+}
 
 export default TitleComponent;
