+++ assignment-6&7/src/Components/index.js
@@ -7,4 +7,5 @@ import ErrorComponent from "./Error/ErrorComponent";
 import TeamMember from "./Team Component/TeamMember";
 import ProfileComponent from "./Profile/ProfileComponent";
 
+
 export { CardComponent, TitleComponent, SearchBar, AboutUs, ErrorComponent, TeamMember, ProfileComponent };
