+++ assignment-6&7/src/Context/ThemeContext.js
@@ -1,0 +1,8 @@
+import { createContext } from "react";
+
+const ThemeContext = createContext({
+    theme: "default",
+    setTheme: () => {}
+})
+
+export default ThemeContext;
