+++ assignment-6&7/src/index.js
@@ -2,6 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import {App, BodyComponent} from './App';
 
+
+
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import { AboutUs, ErrorComponent, TeamMember, ProfileComponent } from './Components';
 
@@ -30,14 +32,12 @@ const AppRouter = createBrowserRouter([
                 ]
             },
         ]
-    }
-    
-    
+    },
 ])
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-    <RouterProvider router={AppRouter} />
+        <RouterProvider router={AppRouter} />
 );
