+++ assignment-6&7/src/utils/usernames.js
@@ -1,0 +1,12 @@
+const usernames = [
+    "akshaymarch7", 
+    "ap221882", 
+    "nikitaj-57", 
+    "Pujarini", 
+    "aditifarkya",
+    "gavandivya",
+    "nitishnivedan",
+    "Bhallora"
+]
+
+export default usernames;
